requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dominate==2.3.5
et-xmlfile==1.0.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
idna==2.8
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10
lxml==4.3.3
MarkupSafe==1.1.1
numpy==1.16.2
openpyxl==2.6.1
pandas==0.24.2
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2018.9
requests==2.21.0
six==1.12.0
soupsieve==1.8
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
xlrd==1.2.0
XlsxWriter==1.1.5
xlwt==1.3.0
